Add files via upload
There is both a student user and an admin user. To begin, login as an admin and enter username Admin and password Admin001. From there, you can initialize as many students as you like and login as students once they have been initialized. Both the admin and student can perform a variety of methods, so feel free to check them out!